+++ srcs/toputinlib1.c
@@ -1,17 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   toputinlib.c                                       :+:      :+:    :+:   */
+/*   toputinlib1.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mclaudel <mclaudel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/12 13:39:00 by mclaudel          #+#    #+#             */
-/*   Updated: 2019/11/13 14:53:43 by mclaudel         ###   ########.fr       */
+/*   Updated: 2019/11/13 16:17:20 by mclaudel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <ft_utils.h>
-#include <stdio.h>
 
 int		incharset(const char *set, char c)
 {
@@ -30,7 +29,7 @@ void	ft_putunsigned_fd(size_t n, int fd)
 	ft_putchar_fd('0' + n % 10, fd);
 }
 
-void		ft_putunsignedbase_fd(size_t n, char *base, size_t len, int fd)
+void		ft_putunsignedbase_fd(unsigned int n, char *base, size_t len, int fd)
 {
 	if (n / len > 0)
 		ft_putunsignedbase_fd(n / len, base, len, fd);
@@ -49,6 +48,8 @@ size_t	ft_leninbase(unsigned int i, char *base)
 		i /= order;
 	return (len ? len : 1);
 }
+
+
 
 int		ft_unsignedlen(unsigned int n)
 {
